feat(transfer + delays) transfers now can have delays.
Trading Agent: NEW METHOD: _get_random_transfer_delay: this is based off of a beta distribution and randomly generates a delay between 5 and 70 mins (variables in init) the peak is around 23mins. This is based off of information that on Binance most transfers are between 10-30 mins, with some going up to 60, and a couple above 60 in a blue moon.

NEW CLASS: CompleteTransferMsg: This class is for self calls, using a delay it allows agents to schedule the completion of a transfer. Such that they are temporarily without holdings for the during of the transfer.

new init: updated to import the message class

core_background_agent: logic updated with the transfer delay call and send_message